move TemplateToken Equals to Sdk
