+++ src/Runner.Language.Server/CodeLensProvider.cs
@@ -16,6 +16,7 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Document;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
 using Runner.Server.Azure.Devops;
+using Sdk.Actions;
 
 namespace Runner.Language.Server;
 
@@ -49,106 +50,6 @@ public class CodeLensProvider : ICodeLensHandler
             Schema = PipelineTemplateSchemaFactory.GetSchema()
         };
         return templateContext;
-    }
-    class Equality : IEqualityComparer<TemplateToken>
-    {
-        public bool PartialMatch { get; set; }
-
-        public bool Equals(TemplateToken x, TemplateToken y)
-        {
-            return TemplateTokenEqual(x, y, PartialMatch);
-        }
-
-        public int GetHashCode(TemplateToken obj)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    private static Exception UnexpectedTemplateTokenType(TemplateToken token) {
-        return new NotSupportedException($"Unexpected {nameof(TemplateToken)} type '{token.Type}'");
-    }
-    private static bool TemplateTokenEqual(TemplateToken token, TemplateToken other, bool partialMatch = false) {
-        switch(token.Type) {
-        case TokenType.Null:
-        case TokenType.Boolean:
-        case TokenType.Number:
-        case TokenType.String:
-            switch(other.Type) {
-            case TokenType.Null:
-            case TokenType.Boolean:
-            case TokenType.Number:
-            case TokenType.String:
-                return EvaluationResult.CreateIntermediateResult(null, token).AbstractEqual(EvaluationResult.CreateIntermediateResult(null, other));
-            case TokenType.Mapping:
-            case TokenType.Sequence:
-                return false;
-            default:
-                throw UnexpectedTemplateTokenType(other);
-            }
-        case TokenType.Mapping:
-            switch(other.Type) {
-            case TokenType.Mapping:
-                break;
-            case TokenType.Null:
-            case TokenType.Boolean:
-            case TokenType.Number:
-            case TokenType.String:
-            case TokenType.Sequence:
-                return false;
-            default:
-                throw UnexpectedTemplateTokenType(other);
-            }
-            var mapping = token as MappingToken;
-            var othermapping = other as MappingToken;
-            if(partialMatch ? mapping.Count < othermapping.Count : mapping.Count != othermapping.Count) {
-                return false;
-            }
-            Dictionary<string, TemplateToken> dictionary = new Dictionary<string, TemplateToken>(StringComparer.OrdinalIgnoreCase);
-            if (mapping.Count > 0)
-            {
-                foreach (var pair in mapping)
-                {
-                    var keyLiteral = pair.Key.AssertString("dictionary context data key");
-                    var key = keyLiteral.Value;
-                    var value = pair.Value;
-                    dictionary.Add(key, value);
-                }
-                foreach (var pair in othermapping)
-                {
-                    var keyLiteral = pair.Key.AssertString("dictionary context data key");
-                    var key = keyLiteral.Value;
-                    var otherv = pair.Value;
-                    TemplateToken value;
-                    if(!dictionary.TryGetValue(key, out value) || !TemplateTokenEqual(value, otherv, partialMatch)) {
-                        return false;
-                    }
-                }
-            }
-            return true;
-
-        case TokenType.Sequence:
-            switch(other.Type) {
-            case TokenType.Sequence:
-                break;
-            case TokenType.Null:
-            case TokenType.Boolean:
-            case TokenType.Number:
-            case TokenType.String:
-            case TokenType.Mapping:
-                return false;
-            default:
-                throw UnexpectedTemplateTokenType(other);
-            }
-            var sequence = token as SequenceToken;
-            var otherseq = other as SequenceToken;
-            if(partialMatch ? sequence.Count < otherseq.Count : sequence.Count != otherseq.Count) {
-                return false;
-            }
-            return (partialMatch ? sequence.Take(otherseq.Count) : sequence).SequenceEqual(otherseq, new Equality() { PartialMatch = partialMatch });
-
-        default:
-            throw UnexpectedTemplateTokenType(token);
-        }
     }
 
     string GetDefaultDisplaySuffix(IEnumerable<string> item) {
@@ -293,7 +194,7 @@ public class CodeLensProvider : ICodeLensHandler
                                                 // The official github actions service reject this matrix, return false would just ignore it
                                                 throw new Exception($"Tried to exclude a matrix key {item.Key} which isn't defined by the matrix");
                                             }
-                                            if (!(matrixexcludeincludelists && val is SequenceToken seq ? seq.Any(t => TemplateTokenEqual(t, item.Value, true)) : TemplateTokenEqual(val, item.Value, true))) {
+                                            if (!(matrixexcludeincludelists && val is SequenceToken seq ? seq.Any(t => t.DeepEquals(item.Value, true)) : val.DeepEquals(item.Value, true))) {
                                                 return false;
                                             }
                                         }
@@ -339,7 +240,7 @@ public class CodeLensProvider : ICodeLensHandler
                                 flatmatrix.ForEach(dict => {
                                     foreach (var item in keys) {
                                         TemplateToken val;
-                                        if (map.TryGetValue(item, out val) && !TemplateTokenEqual(dict[item], val, true)) {
+                                        if (map.TryGetValue(item, out val) && !dict[item].DeepEquals(val, true)) {
                                             return;
                                         }
                                     }
